backend/requirements.txt
Django==4.1.4
gunicorn==20.1.0
matplotlib==3.7.1
numpy==1.24.2
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
randimage==1.2
whitenoise==6.3.0
channels==4.0.0
channels-redis==4.0.0
daphne==4.0.0
django-cors-headers==3.13.0
django-storages==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
boto3==1.26.88
requests==2.28.2
django-fieldsignals==0.7.0
# torch==2.0.0